refactor(core): 重构核心模块并迁移图表相关工具函数
- 将 chart-tools、chart-data 等工具函数从 lib 移动到 core/utils 目录
- 更新相关文件的导入路径
- 删除 lib 目录下的冗余文件
- 优化部分代码结构以提高可维护性
