+++ src/core/types/chart-type.types.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import type {
   CandlestickData,
   HistogramData,
