+++ src/core/utils/time.utils.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import type { TimeframeSec, TimeframeStr } from '../types/time.types';
 
 /**
